requirements.txt
langchain-google-genai
python-dotenv
Pillow
pygments
gTTS
pyttsx3
pydub
moviepy